third_party/blink/web_tests/platform/win/external/wpt/resource-timing/cross-origin-start-end-time-with-redirects-expected.txt

This is a testharness.js-based test.
FAIL Verify that cross-origin resources don't implicitly expose their redirect timings assert_less_than: startTime should not expose redirect delays expected a number less than 2000 but got 2012
Harness: the test ran to completion.
